+++ doc/source/changelog.rst
@@ -14,6 +14,7 @@ master: 2.4.0
 -------------
 
 * Support for globally overloaded ordering operators
+* Special cases for __repr__/__str__ returning C++ stringy types
 * Fix lookup of templates of function with template args